Fix: persistent_workers crash, Lovász fp32, sampler seed accumulation
1. Remove persistent_workers=True — causes "Trying to resize storage
   that is not resizable" when workers cache tensor storage
2. Lovász loss now computes in float32 (AMP was giving float16,
   accumulation over B samples lost precision)
3. ClassBalancedTileSampler.set_epoch uses base_seed + epoch
   instead of accumulating (seed was drifting non-linearly)

+++ simple_losses.py
@@ -59,8 +59,12 @@ class LovaszLoss(nn.Module):
         if logits.dim() == 4:
             logits = logits.squeeze(1)
 
+        # Compute in float32 for numerical stability (AMP может дать float16)
+        logits = logits.float()
+        targets = targets.float()
+
         B = logits.shape[0]
-        loss = torch.tensor(0.0, device=logits.device, dtype=logits.dtype)
+        loss = torch.tensor(0.0, device=logits.device, dtype=torch.float32)
 
         for i in range(B):
             flat_log = logits[i].reshape(-1)
